C# Executable to launch ENBInjector.exe and then fnv4gb.exe. The Injector will be automatically closed when Fallout NV is closed.
C# Executable to launch ENBInjector.exe and then fnv4gb.exe. The
Injector will be automatically closed when Fallout NV is closed.

I created this quick console applicaton when I started using an enb and
saw that two executables needed to be launched to run Fallout New Vegas
with an ENB and a modern system. I created this application in c#.